requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
flask==2.3.3
jinja2==3.1.2
aiohttp==3.9.1
flask-cors==5.0.1
gunicorn==21.2.0
